address comments for clarity
